V0-Livro

+++ src/main/java/com/kata/books/repositories/LivroRepository.java
@@ -3,4 +3,4 @@ package com.kata.books.repositories;
 import com.kata.books.domain.Livro;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LivroRepository extends JpaRepository<Livro, Long> {}
+public interface LivroRepository extends JpaRepository<Livro, Integer> {}
